requirements.txt
alpaca-py>=0.30.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pytz>=2024.1
